backend/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.1.1
sqlparse==0.5.3
